requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
h5py>=3.8.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
pre-commit>=3.5.0
commitizen>=3.13.0
tqdm>=4.66.0
pyyaml>=6.0
